Z and sigma 2d icestep with P1dg lim with SquTri on tracers and vel
Using squeezed triangles filter (from last commit) with P1dg limiter
on tracers on a z layer mesh still crashes around 30 days (when
density front propagates up ice front). There is an
improvement over the limiter without the limiter in the initial
spinup.

However on a sigma layer mesh the above is stable to 100 days. The
initial spinup with/wout the limiter is identical until the density
front reaches the ice front. With the limiter switched on an
instability/wave is reflected back into the cavity so the melt rate
has spurious oscillations. These are similar with/wout squeezed
triangles switched on.

Using a different melt parameterisation which does not depend on velocity
on the z layer mesh leads to a stable simulation. However the melt rate
calculated using a limiter on the tracers is consistently higher.

Adding a limiter to a DQ1 velocity component and using this in the melt
rate calculation and scalar advection also makes the simulation stable
to 100 days. Again the melt rate is consistently higher with the limiter on.
When you look back at z layer simulations with the limiter only on the
tracers (i.e the one that was crashing at 30 days above) the limiter
is also making the melt rate too high cf without limiter.
Also, using the velocity just in the melt calculation (i.e not tracer
advection) gives similar results to usign the limited velocity in both
the melt calculation and the scalar advection.

The plan is to test if solution changes by varying timestep/grid resolution.
Ultimately sigma layer model is showing that limiter shouldn't be needed
under ice shelf if the dynamics are being resolved smoothly.

Separately a change in the grid at the ice front might be needed.
More than two stacked grid cells at front might cause problems with
top cell. need to to see if unstable spurious density gradients
appearing...