Change defaults of u,p boundary conditions:
Boundaries that are not specified are now assumed closed, i.e. u.n=0.

The scalar equation should work the same as before, *except* on
boundaries that are not included in ds (the vertical boundary facets in
the interior of the domain) it now does the same things as on all other
boundaries.