Simulation still crashes only applying dirichlet in div term
This was last (?) attempt to try and get vertical ice front to work.

With the momentum advection change in the last commit the simulation
was more stable at higher Kv (1e-2m^2/s) actually getting to 100 days
and surviving the first density front leaving the cavity. However, it
still crashed with Kv =1e-3. If you added the Kv switch in (increase
Kv to 0.1m^2/s if unstable density gradient) then the simulation was
stable. Good news was that this only kicked after the density front
was past the ice shelf. Before with with old limiter the vertical density
gradient field went positive (i.e unstable) during the spinup, when it
probably should not have done.

All that was good, but then if you increase the horizontal resolution
the simulation crashed much earlier at 7 days, long before the density front
had got to the ice front. The blow up still started at the ice front
though.

Obviously something wrong with the bcs setup.

I had a long meeting on Friday evening with Stephan talking about bcs.
He reckoned that viscosity/diffusivity terms were fine. Last thing to
try was if we should only apply the dirichlet u.n =0 condition on the
divergence term. In momentum advection technically the dirichlet condition
sets u = 0 so you are affecting the tangential flow too. A key part
of this is that you are trying to preserve the symmetry of the the divergence
term with the grad pressure term. That is why appyling dirichlet condition
on the velocity enforces a neumann condition on the pressure.

I added a label key 'closed' as part of the dictionary so that it would
only apply u.n if 'closed' was not in th dictionary. However the simulation
still blew up at the same time.

I think we are going to move on from vertical ice front now. We may have to
come back if the squeezed triangle limiter doesnt work in 3d...