Typo fix seconds in year and increase default velocity
Corrected typo for the number of seconds in a year.

Increase the default velocity so that the calculated
gammaT = GammaT * ustar ends up closer to the default
constant value gammaT. This reduces the percentage errors
output at the end of the script. Emphasises that the melt
scheme is very sensitive to velocity (ustar) at the boundary
and the resultant value of gammaT...