Stable 3d with higher penalty term for viscosity
Tracked down the odd behaviour (pressure increasing by initial pressure
each time - see last commit) to the viscosity term.

The momentum solves were completely failing to converge. (true residual
actually going up after an iteration). In the initialisation of pressure,
when you have an explicit solve the only nonzero term is due to buoyancy, so the
predicted vertical velocity is very high. This was the first guess
for each subsequent momentum solve. Since the solver thought it was converging
(because the preconditioned residual was so large and then dropped arbitarly below
the relative tol criteria) each time the pressure correction step was adding on the
initial pressure to balance the high vertical velocity coming out of the momentum solve.

Increasing the penalty term made the simulation stable and runs to 100 days ok.

The problem was probably the polynomial degree fed into the ip scheme was too low.
Our default was 'degree-1', which would be the degree of a gradient field
on usual triangles/tets. But because we are using extruded meshes with tensor
products of the elements the degree of the gradient field actually comes out as
the degree of the normal field (2).

horizontal velocity is a product of N2curl(1) (which is a really a subset of P1 ^2) and
P2. Stephan and I talked about this logic for the degree of the gradient field:
- horizontal gradient: P0 x P2 -> degree = 2
- vertical gradient: N2curl(1) x P1 -> degree = 2

vertical velocity is a product of P2 x P1DG
- horizontal gradient: P1 x P1DG -> degree = 2
- vertical gradient: P2 x P0 -> degree = 2

Anyway by itself that change wasn't enough, as the ip factor had to be set to at least 2.5
(before the default was 2). At the moment the new degree is hard coded in but probably this
should have an option depending on the type of discretisation.

The number of iterations for the momentum solve in the first 20 steps is a bit sensitive to
the ip factor. i.e with ip = 3 there were 20 fewer iterations required (50ish) compared
with ip = 4. (70ish)

Loosening the tolerance of the momentum solve from 1e-7 to 1e-5 also led to a drop in iterations,
from 55 to 25 after 20 timesteps. Bu there was no obvious visual increase in the noise level.