Resetting u* after pressure init reduces momentum solve iterations and noise
Linked to the previous commit when we noticed that resetting u* after the pressure
initialisation to uold (i.e the zero vector currently) led to a reduction in the
true residual before the solve, although didn't fix the underlying problem (the viscosity
term).

After sorting out the viscosity term (previous commit) we found that resetting u* after the
initialise_pressure routine led to a reduction of 30 iterations in the second momentum step
(before: 103, after: 73).

It also led to a visually obvious drop in the noise around the grounding line. Presumably because
it removes the high vertical velocity signal from the initial momentum solve (balancing the
buouyancy term).

Resetting u* in subsequent 'real' momentum solves had no noticeable affect on the noise
or number of iterations however.