Offsetting melt bcs prevents spurious freezing at ice front in 2d
Offsetting the melt bcs (temperature and salinity Neumann bcs) by 5 grid
cells away from the ice front transition prevents the spurious freezing
we saw before when refining the resolution without a limiter.

This did not carry through to 3d, however, as a cross section through the
interior of a 100 day 3d simulation showed that there was freezing at the ice
front transition.

So probably the solution is to have a 3d 'squeezed wedge' filter for the
P1dg limiter to minimise the effect of the p1dg filter in the squeezed wedges
at the ice front...