start ekman 2.5d test/demo for extruded meshes
For some reason 2.5d no longer works for extruded meshes on the
extrusion support branch.

In November 2020 I had a 2.5d setup for isomip+ on an extruded
mesh which worked ok, but now doesn't. (I never added it to
github partly because it used the old/wrong DQ1-Q2 element pair.

There is a key error when the timestepper is set up for the
u velocity...

I also tried the old tets ekman demo from December 2019. This gives
a similar error on the extrusion support branch but not on
the master branch...