P1dg limiter prevents freezing at ice front transition in 3d
Tracer fields were becoming unbounded in the 3d isomip+ setup
specifically in the cell on the 'left' side at the ice front
transition. i.e temperatures down to -6degC. Presumably this is
because coriolis is causing fast flow on the left hand side.
There is the acceleration at the ice front transition which leads
to very fast flow speeds. This causes the melt parameterisation
to do lots of melting, which drastically lowers the temperature,
ultimately leading to freezing conditions. Since this all occurs
over one cell, it is definitely not resolved.

Turning on the limiter fixes the problem. i.e after 60 days
the min temperature field is -2.01degC. This is good agreement
with the colour scale used in the isomip+ paper -2.4degC - 0.8degC

In the first few days there is the noticeable 'leaking' from the
squeezed wedges, similar to what we saw in 2d. Ultimately though
at 60 days these features are no longer obvious compared to the
main flow speeds (max x vel = 0.6m/s)

Also on the plus side if the flow speeds are ~ 0.1m/s then as the domain is
480km long it will take 480e3 / 0.1 = 4.8e6s for a signal to
propagate from the grounding line to the edge of the box.
This is 55 days.