Mesh ISOMIP+ domain with unstructured base mesh
This should have been a few commits, but hopefully
includes all the changes needed for the ISOMIP+
domain with bathymetry / ice draft.

The base mesh was made first. QGIS extracted a contour of
10m ocean cavity thickness and then qmesh (using docker), and
qmesh-cli from bitbucket, could process this to produce a mesh.

Jon Hill made some good demo videos on youtube and I wrote a
few notes on Teams, but I do need to combine them all.

The first go at the mesh produced very noisy results at the ice
front because the triangles weren't aligned along the ice front
so the slope which was smooth before became very 'jagged'.

The trick was to draw two interior (i.e not joined to the outer edges)
 lines either side of where the ice front transition was supposed
to take place so that the triangles became aligned.

The current configuration is hardcoded to 4km horizontal resolution
(from generating the base mesh) so the real thing needs to be
2km res. But all in all it seems stable.

The iteration counts for pressure were pretty identical.
I didnt have a like for like comparison for the momentum iterations.
At 100 days mom iterations were 58. which was more than before on structured
mesh despite now having fewer dofs (1344266 dofs cf 1587232 before).

(in the next commit changing to crank nicholson and reducing the timestep
reduces the momentum iterations and the pressure substantially,
though it does get slower....)