Crank Nicholson
Try crank nicholson (theta = 0.5) compared with default backward euler
 for momentum timestepping.

I was wondering if maybe backward euler was too diffusive because
we dont see horizontal eddies that Paul was expecting...

USing same timestep of 900s crank nicholson looks similar to backward
euler except on the western wall which has significant stripes. The max vel
has also shot up to 2.9m/s

Mom iterations did drop though from 58 with BE to 29 with CN. Pressure dropped
from 78 BE to 68CN. U correction and tracers pretty much the same.

With dt = 450s stripes were reduced but still quite obvious and max vel came down
. At 70 days  iterations with 450s: mom= 26, press = 61, ucorr = 12, tracers = 10.

And then tried with dt = 225. Again the stripes are reduced compared to
dt = 450s. Especially on the inside so that in the interior no obvious stripes.
The max velocity is also down to ~ 0.7m/s which was consistent with BE before.
 at 15 days mom = 21, ucorr = 11, press = 52,
tracers = 8.

So iterations counts are going down, but reducing the
timestep does increase runtime! with dt = 225 on 8 cores on my machine
it takes 6hours 20mins to run 1 day of simulation! (1:3.8) (dt = 450s was
1:7)

I stopped dt = 225s at 15 days because I want to try and sort out the
laplace preconditioner to work with open bcs.

Ultimately we need this to work on a finer mesh with dx = 2km...
Maybe the velocity function space is too diffusive...

(ALSO I am not sure if I definitely need to change theta to theta_p
in coupled integrators. This was something that prevented excess noise
in the velocity before in February 2021, but I didnt check it that carefully)