Fix restoring stress, plus outer ksp required
Setup a 2d script to have open boundaries. Slightly annoying that
I obviously left this half way because the domain shape is bonkers
with a really tall grounding wall...

Fixed a problem with the restoring stress, as I had forgotten to
subtract off the reference temperature and salinity.

Running with mumps for the coupled correction solve (i.e for the
whole block system without doing the schur factorisation) seems to
work ok and I've used this simulation as the reference 'best' solution.

The previous default solver settings (for extruded meshes) uses
the laplacePC in place of the full schur complement (to avoid
having to solve a mass matrix inversion for every cg iteration, n.b
the matrix system is G^TM^-1G p = f)

This does not seem to work before (as Stephan suspected) because the
velocity solution is quite a bit more noisy. Instead if you activate
the outer krylov solve for the Fieldsplit 1 block everythign is fine,
and the solution looks like mumps.

There are two ways to do this. One is to restrict the laplace pc step to
only be a preconditoner. I think this means just do a multigrid step and
then pass the solution back up to the outer cg loop. In the 100th step
(seems to have converged to steady behaviour by this point)
this leads to 29 steps of the outer cg loop. But importantly each of these
steps has the inner laplace / lumping preconditioenr and a mass matrix solve,
which takes about 6-7 iterations.

But if you allow the laplace step to also be a krylov solve then in the final
step, there are 29 iterations of the inner laplace solve, with only ONE of
the outer krylov solves. This means there is only one extra mass matrix solve
as opposed to 29 before.

I tried relaxing the tolerance of the inner laplace solve in case there was a
tradeoff in having more outer solves vs fewer inner solves. But this didnt seem
to work. Instead the outer krylov solve always converged in one solve and the
inner solves just hadn't reduced the residual as much. There was more noise in the
solution when this happened.

If you add a (ksp) richardson outer outer (!) solve i.e just repeat the coupled solve
until the residual goes down this makes sure the noise goes away but it ends up
being more expensive because it forces multiple inner solves to happen. I.e with
tol 1e-4 I think you ended up with 29 + 5 iteratiosn of cg due to the extra
richardson solve...

I also tried getting rid of the laplace (and vertical lumping completely) and using
sor / gamg as the preconditioner for the schur complement, but this didnt seem to be
converging at all. I am kind of surprised that gamg didnt work so maybe I just
didnt get the option right...

Anyway the proof of the pudding will be in 3d...

Finally I think the weak form for the surface boundary conditions is correct
because turning combinations of the extra stuff I added off makes the
solve diverge.