Setup isomip+ 2d domain and first inflow tests
This is a bit messy so need to tidy up later.

Had to allow 1d mesh to be extruded in utility.py
and then had to get the right format for interpolate for a 1d line.
Probably just tidy it up with an if statement which checks the
geometry dimension of the input function.

Then in the runscript there was a bit of tweaking to get something
stable. Initially the limiter was off so maybe increasing the salinity
to 20PSU for the inflow water is not necessary. With the limiter turned off
there is a sinking density plume which is unstable unless the horizontal
viscosity is doubled. I'm not sure yet if this is a 'real' feature we
want to preserve or not... The limiter smoothes this feature out.