Add adjoint back in... no init pressure... and still direct (mumps)
Added adjoint changes back in slowly!

Reading from dump taylor tests don't seem to converge. Maybe because
vertical resolution is too coarse? Taylor test are better with
30 layers cf 15... (dx =4km, dt = 900)

Using conditional in the melt objective function works fine until
tape.eval_adj() is turned back on. Then you get an obscure adjoint error.
Using the smooth backwards step instead doesnt raise the error
and seems to work fine. Plus using the smooth backwards step in temp/sal
bcs seems to work better than conditional in a taylor test after 1 step.
(But this might have been a resolution dependency thing)

So then with eval_adj() but with the backwards step functional
you get an error about a discontinuous facet integral needs to be
restricted, which is all about interior facets. This seems to be
to do with the initial pressure solve because if you comment this out
then it runs ok... I suspect that changing back to iterative solvers
and not having an initial pressure solve will cause a problem...

I've done taylor tests after ten steps with temp, mu, kappaT and kappaS
as the controls. All looks good. You have to be a bit careful with
the magnitude of h for mu / kappa. Mu the best result is with
0.01 -> 1.97, 1.979, 1.971. With Kappa (sal) the best result is with
1,98, 1.991, 1.9944, which happens with h 1e-3. (Both of these are after
ten steps)