Make scalar solves iterative
In changing the solves to iterative I noticed that the pressure nullspace
was missing.
Adding it back in with all iterative solvers causes the taylor tests to break
i.e convergence only order 1. And the forward pressure solution is noticeably
different as are the adjoint fields near the grounding line. Velocity and
scalars seem the same though.

Using mumps on the momentum + pressure correction solve (and iterative on
the scalar solves) with the nullspace seems to be ok. Taylor tests convergence
back to 2. The pressure looks the same as 'new pressure' calcualted with
the iterative solver after the nullspace is back in. But the adjoint
fields are identical to the 'old'values before the nullspace is added in.
So possibly this is just a tolerance issue? Or maybe it is because
we haven't got the initial pressure dump...