Add more adjoint tests
Add more failing and passing 2d adjoint tests based on the isomip+
setup.

Turning on tape.evaluate_adj() causes two problems:
1) First an error occurs related to the contional in the melt functional.
This can be removed by using the backward step.

2) Then an error occurs if you do multiple timesteps which seems to
be due to the initial pressure solve. The error is about integrating over
discontinuous interior facets. N.b removing the pressure solve only works
if you use mumps for the coupled solves though as the iterative solvers
need the initial solve to have a good first initial guess.

I thought (2) wouldnt be a problem because we are loading from a dump file
anyway with an initial pressure. I've tested with a dump after 100 days of simulation
The taylor tests are fine if the control is GammaT but using the initial
salinity as a control the taylor tests aren't converging at all...