Pressure still not converging.
Changed back to nonzero pressure with gradp.n = 0 on all boundaries.

Using dt = 100s (from 10 percent u / p off analytical) get 2.34 pressure convergence
on grids 10 ->20, but not after to 80... (40-> 80: 0.79). The finer grids
get to the stopping criteria difference much earlier low. If you force the
velocity stopping criteria down to 1e-8 you can get 1.46 but it still isnt converging.

If you make the stopping criteria tighter so more timesteps are run the reason the convergence
looks better seems to be because the coarser grid error goes up, rather than
 the finer grid error going down. This seems odd. Maybe the initial start
just is not working so it does need to spinup from zero? I haven't got this
to work yet because of robustness but maybe that is the next thing to try...