New rasterio with transformed bathymetry in new tiff
Before the old version of rasterio (<=1.2.6) was able to read
the isomip+ netcdf file and automatically do the transforming
to get the right coordinates. I think in rasterio 1.2.7 there was
a major (?) change to follow gdal to be tighter on the georeferencing
and because this netcdf does not have a 'transform' or 'crs' I think
the new default behaviour was to not modify it. In the end the easiest
thing was to just write another file with the correct transform
as a tif file. This seems to work fine on ese-wiscott but there is still
an issue (maybe?) on ese-floyd... (a new computer January 2023...)