hack for 3d isomip plus bathy to work
The next problem is that running 3d isomip plus with the new rasterio
changes doesn't work with the latest version of scip (1.10.0). It works
fine in 2d but there seems to be a problem with a memory buffer... see
issue #4 for the trace back.

It seems like it actually works with an older version of scipy 1.9.1
so the hack for the moment is just to force the user to use scipy
1.9.1 if they do pip install -r requirements.txt as part of the setup.