Add Trestore and Srestore as adjoint variables
I realised that I didn't have T/S restore as dirichlet bc. The sponge
region was the only thing having an effect. If you add dirichlet conditions
the forward model at the rhs adjusts from the dump to a stratification pattern
where the salinity and temperature 'tilt' up in order to satisfy the bc with
T=Trestore everywhere along the boundary (as opposed to just relaxing to that value
with the sponge region). Also the simulation is closed. looking at these things
made me rethink the thwaites/icefin ground zone simulations at the beginning
of February. One thing that came out of that was trying to have 'qadv' bc where
the dirichlet value is only used on the inflow for open bcs... maybe I could try doing
this?