First attempt at running inversion
run two forward simulations: 1) constant init/rhs T and 2) linear init
/rhs TS. vel and melt very similar but slightly different Ts stratification.
linear TS has to curve to fit dS/dz = 0 at the bottom. Also this
was changed from original ice shelf coarse.py because this has open boundary conditions
and 'qadv' bc on rhs so the stratification is aligned with the ice ocean interface
all the way to the rhs (there are no funny boundary effects for the constant
TS case). It also uses the dynamic pressure bcs.

added objective function based difference between two simulations after 20 days.
taylor tests are good for all values except Srestore which seemed maybe after
adding weights but not obviously right... Pushed ahead with inversion....

Srestore and Trestore fields are changing only at the boundary but not
converging to right answer. regularisation is 'wrong' in that it is
penalising gradients but the 'target' solution has a linear gradient
in the vertical... 0.1*reg seemed better than reg anyway... This was
'only' running the inversion forward for 10 days. based on the forward runs it might
take longer for the signal to propagate through from the beginning...