Add 'qadv' and float options...
qadv allows you to keep homogenous neumann as default bc but apply
a dirichlet value on inflow bcs... this seems to be good for melting
so that the open boundary has a less obvious impact on the interior
solution.

'float' option was a test to see if just letting the neumann bc 'float'
i.e solve as part of the sytem might help... this did not seem to have any effect...
will probably remove in a later commit.