Try inversion for 4 params: slope/intercep init/bc
Quite a few changes from the last commit.
Using the restorefield for T and S didnt work for the inversion
likely because tape not working properly through scipy...

Added bounds, particuarly for the slopes to stop getting crazy values
for T and S. Specifically, for the moment, S can't have a positive
(unstable) slope but T can because relatively it makes less difference
to the density.

So far don't know if this works but need to have a basis to keep moving
from. I have tried trying to invert from a dump file but this
seemed to be unstable probably due to picking up the old velocity /
TS fields that are then immediately inconsistent with the right hand
side values coming from the optimisaiton. Now I have switched to having
the simulation start from rest which seems to at least allow the
optimisation algorithm to get past the first step... I'm running the
forward model for 1 day and 10 days (with dt = 300s and muh = 0.25m^2/s
like before) hoping that 10 days should be long enough for the signal to
propagate from the rhs to the depth profile...

added some callbacks (and a adjoint diagnostic for constants).

Note also I should have commited somewhere that using P1DG for the old
sensitivity to restore field ensure that sensitivity was only in the
final cell (because using L2 for sensitivity involves multiplying
by a mass matrix) and for P1 you get 'leaky' sensitivites in cells
not directly connnected to the rhs boundary.