Adjoint setups with L2 boundary norms
ice_shelf_coarse.py now has GammaT as a control - this required changing
the melt parameterisation to the Jenkins2010/ISOMIP formulation.

ice_shelf_coarse_open_balp.py is a new file with the depth profile
objective function for sensitivies only.

I initially tried with the new balance pressure code solving for
the pressure that balances density/geostrophy but this was causing
undulations at the rhs. Maybe this is because the sloped domain
has isopycnals that are not flat at the boundary but the balance pressure
solve was enforcing zero at the boundary... not sure..