Don't doubly apply bcs on outflow
After chatting with Stephan if specifying a bc id don't add the
boundary integral twice on outflows!

There is a slight subtletly with the previous else option fo specifying
uin. It only set a u value of zero in the momentum advection term. It doesnt
set u in the continuity equation. This might be useful as a stabilising
term (for instance in the overturning.py script with the open top) but
it is not obvious that it makes physical sense because if you specify a
stress bc then the velocity can be nonzero